add mouse_hooks.c
