+++ inc/cub3d.h
@@ -40,9 +40,14 @@ int		game_start(t_all *game);
 /*raycast*/
 void	draw_location(t_all *game);
 void	init_raycast(t_all *game, int x);
-/*util*/
+
+/*hooks*/
 int		key_hook(int key, t_all *game);
+int     mouse_hook(int x, int y, t_all *all);
 int		exit_hook(void);
+void	key_hook_lr(int key, t_all *game);
+
+/*util*/
 void	ft_error(char *str);
 char	*del_n(char *line);
 void	my_pixel_put(t_all *game, int x, int y, int color);
