+++ inc/cub3d_defines.h
@@ -25,11 +25,11 @@
 # define KEY_ESC			65307
 # define KEY_SPACE			32
 # define KEY_ENTER			65293
-# define W_WIDTH 1920 // 1280/1920
-# define W_HEIGHT 1080 // 720/1080
+# define W_WIDTH 1280 // 1280/1920
+# define W_HEIGHT 720 // 720/1080
 # define T_WIDTH 64
 # define T_HEIGHT 64
-# define RS 0.3
+# define RS 0.20
 # define MS 0.25
 
 #endif
