+++ maps/map_b.cub
@@ -1,7 +1,7 @@
-NO ../xpm/no.xpm
-SO ../xpm/so.xpm
-WE ../xpm/we.xpm
-EA ../xpm/ea.xpm
+NO ./xpm/no.xpm
+SO ./xpm/so.xpm
+WE ./xpm/we.xpm
+EA ./xpm/ea.xpm
 
 F 220,100,0
 C 225,30,0
